Move index.js XRIFrame implementation to separate file
